Restructure app
Organize code around responsibilities
