+++ src/app.module.ts
@@ -1,26 +1,7 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { CountriesService } from './countries.service';
-import {
-  CountryListProviderToken,
-  LandborderListProviderToken,
-} from './dataset/constants';
-import countries from './dataset/countries';
-import landBorders from './dataset/landborders';
+import { RestApiModule } from './rest-api/rest-api.module';
 
 @Module({
-  imports: [],
-  controllers: [AppController],
-  providers: [
-    CountriesService,
-    {
-      provide: CountryListProviderToken,
-      useValue: countries,
-    },
-    {
-      provide: LandborderListProviderToken,
-      useValue: landBorders,
-    },
-  ],
+  imports: [RestApiModule],
 })
 export class AppModule {}
